Update acquisition cost type and API route handling
This commit includes alterations to the types used for acquisition costs in both the "fixed_assets" and "formulas" tables. The changes reflect the need for greater precision, hence the shift from BigInteger to double types.

In the API routes, the 'getIP' route is now accessible without authentication, which enables better handling of unauthenticated printer IP requests. This change was also duplicated within the 'auth:sanctum' route group.

The API throttling limit was raised from the default limit to cope with increases in request volume.

In FixedAssetImportController, the response no longer includes 'data', in adherence to a shift in response format standards.

In MasterlistImport, a typographical error in the 'tag_number' field has been corrected. Commented code blocks have been left for possible future use.

In Kernel.php, an unnecessary whitespace was removed for better code cleanliness. Other negligible changes include additions of helper comments for exception handling.

+++ database/migrations/2023_04_27_091123_create_formulas_table.php
@@ -18,15 +18,15 @@ class CreateFormulasTable extends Migration
             $table->string('depreciation_method');
 //            $table->decimal('est_useful_life' , 10, 1);
             $table->date('acquisition_date');
-            $table->bigInteger('acquisition_cost');
-            $table->integer('scrap_value');
-            $table->bigInteger('depreciable_basis');
-            $table->bigInteger('accumulated_cost');
+            $table->double('acquisition_cost');
+            $table->double('scrap_value');
+            $table->double('depreciable_basis');
+            $table->double('accumulated_cost');
             $table->integer('months_depreciated'); //months
             $table->string('end_depreciation')->nullable(); //date format yyyy-mm
-            $table->BigInteger('depreciation_per_year');
-            $table->BigInteger('depreciation_per_month');
-            $table->BigInteger('remaining_book_value');
+            $table->double('depreciation_per_year');
+            $table->double('depreciation_per_month');
+            $table->double('remaining_book_value');
             $table->string('release_date')->nullable(); //date format yyyy-mm-dd
             $table->string('start_depreciation')->nullable(); //date format yyyy-mm
             $table->softDeletes();
